pyproject.toml
[project]
name = "ai-ideago"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.115.11",
    "greenlet>=3.1.1",
    "groq>=0.19.0",
    "langchain>=0.3.20",
    "langchain-groq>=0.2.5",
    "passlib>=1.7.4",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "python-slugify>=8.0.4",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
]
